+++ gui/slick/js/configNotifications.js
@@ -352,36 +352,20 @@ $(document).ready(function(){
 			});
 	});
 
-	$('#testTrakt').click(function () {
-		var trakt_api = $.trim($('#trakt_api').val());
-		var trakt_username = $.trim($('#trakt_username').val());
-		var trakt_password = $.trim($('#trakt_password').val());
-		if (!trakt_api || !trakt_username || !trakt_password) {
+	$('#pinTrakt').click(function () {
+		var trakt_pin = $.trim($('#trakt_pin').val());
+		if (!trakt_pin) {
 			$('#testTrakt-result').html('Please fill out the necessary fields above.');
-			if (!trakt_api) {
-				$('#trakt_api').addClass('warning');
-			} else {
-				$('#trakt_api').removeClass('warning');
-			}
-			if (!trakt_username) {
-				$('#trakt_username').addClass('warning');
-			} else {
-				$('#trakt_username').removeClass('warning');
-			}
-			if (!trakt_password) {
-				$('#trakt_password').addClass('warning');
-			} else {
-				$('#trakt_password').removeClass('warning');
-			}
+			$('#trakt_pin').addClass('warning');
 			return;
 		}
-		$('#trakt_api,#trakt_username,#trakt_password').removeClass('warning');
+		$('#trakt_pin').removeClass('warning');
 		$(this).prop('disabled', true);
 		$('#testTrakt-result').html(loading);
-		$.get(sbRoot + '/home/testTrakt', {'api': trakt_api, 'username': trakt_username, 'password': trakt_password})
+		$.get(sbRoot + '/home/pinTrakt', {'pin': trakt_pin})
 			.done(function (data) {
-				$('#testTrakt-result').html(data);
-				$('#testTrakt').prop('disabled', false);
+				$('#pinTrakt-result').html(data);
+				$('#pinTrakt').prop('disabled', false);
 			});
 	});
 
@@ -568,4 +552,10 @@ $(document).ready(function(){
 		}
 	});
 	if ($('input[id="use_plex"]').is(':checked')) {$('.plexinfo').removeClass('hide')}
+
+	$('#testTrakt-result').html(loading);
+	$.get(sbRoot + '/home/get_connected_Trakt_Account')
+		.done(function (data) {
+			$('#pinTrakt-result').html(data);
+		});
 });